最低限使用可能な見た目に整えるまで

+++ app/Http/Controllers/PostController.php
@@ -56,23 +56,31 @@ class PostController extends Controller
         return view('posts.show')->with(['post' => $post])
         ->with(['replies' => $post->getReply()]);
         
-        else 
-        return view('posts.show_reply')->with(['post' => $post])
-        ->with(['replies' => $post->getReply()])->with(['source' => $post->getSource()]);
+        else{
+            $parentPost = Post::find($post->post_id);
+            return view('posts.show_reply')->with(['post' => $post])
+                ->with(['replies' => $post->getReply()])->with(['parent' => $parentPost]);
+        }
+
     }
     
     public function edit(Post $post)
     {
-        
         if($post->user_id==Auth::id()){
-            $tag_input="";
-            $tags=Post_Tag::where("post_id", $post->id)->get();
-            foreach($tags as $tag)
-            {
-                $tag_name=Tag::where("id", $tag->tag_id)->first();
-                $tag_input.="#".$tag_name->name;
+            if($post->post_id==NULL){
+                $tag_input="";
+                $tags=Post_Tag::where("post_id", $post->id)->get();
+                foreach($tags as $tag)
+                {
+                    $tag_name=Tag::where("id", $tag->tag_id)->first();
+                    $tag_input.="#".$tag_name->name;
+                }
+                return view('posts.edit')->with(['post' => $post])->with(['tag' => $tag_input]);
             }
-            return view('posts.edit')->with(['post' => $post])->with(['tag' => $tag_input]);
+            else{
+                $parentPost = Post::find($post->post_id);
+                return view('posts.edit_reply')->with(['post' => $post])->with(['parent' => $parentPost]);
+            }
         }
 
     }
@@ -95,19 +103,30 @@ class PostController extends Controller
     
     public function update(PostRequest $request, Post $post)
     {
-        $input_post = $request['post'];
-        $post->fill($input_post)->save();
-        $tag_old=Post_Tag::where('post_id', $post->id)->delete(); //元々あったタグを削除
-        preg_match_all('/#([a-zA-Z0-9０-９ぁ-んァ-ヶー一-龠]+)/u', $request->tag_name, $match); //#表記されたタグを配列に格納
-        foreach($match[1] as $input_tag)
-        {
-	        $tag=Tag::firstOrCreate(['name'=>$input_tag]);//tagテーブルにないときは新しくtagを作成
-	        $tag=null;
-            $tag_id=Tag::where('name',$input_tag)->get(['id']);
-            $post->tags()->attach($tag_id);
+        if($post->user_id==Auth::id()){
+            if($post->post_id==NULL){
+                $input_post = $request['post'];
+                $post->fill($input_post)->save();
+                $tag_old=Post_Tag::where('post_id', $post->id)->delete(); //元々あったタグを削除
+                preg_match_all('/#([a-zA-Z0-9０-９ぁ-んァ-ヶー一-龠]+)/u', $request->tag_name, $match); //#表記されたタグを配列に格納
+                foreach($match[1] as $input_tag)
+                {
+        	        $tag=Tag::firstOrCreate(['name'=>$input_tag]);//tagテーブルにないときは新しくtagを作成
+        	        $tag=null;
+                    $tag_id=Tag::where('name',$input_tag)->get(['id']);
+                    $post->tags()->attach($tag_id);
+                }
+                return redirect('/posts/' . $post->id);
+            }
+            else{
+                $input_post = $request['post'];
+                $post->fill($input_post)->save();
+                return redirect('/posts/' . $post->id);
+            }
         }
+            
+                
         
-        return redirect('/posts/' . $post->id);
     }
         public function reply(Post $post)
     {
